Final-Prototype/Project-Demo/requirements.txt
asgiref==3.10.0
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.7
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
shap==0.50.0
six==1.17.0
slicer==0.0.8
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
